+++ presidio_evaluator/entity_mapping/definitions.py
@@ -87,6 +87,7 @@ HIERARCHY: dict = {
             "SOCIAL_HANDLE": ["QQ"],  # QQ: Chinese messaging platform ID
         },
         "LOCATION": {
+            "_aliases": ["LOC"],
             "ADDRESS": {
                 "STREET_ADDRESS": [
                     "STREET",
@@ -129,10 +130,10 @@ HIERARCHY: dict = {
             ],
             "LOCATION_OTHER": ["LOCATION-OTHER", "ORDINALDIRECTION"],
             "GPE": ["GLOBAL_POLITICAL_ENTITY"],
-            "LOC": [],
             "GEO": [],
         },
         "ORGANIZATION": {
+            "_aliases": ["ORG"],
             "COMPANY": [
                 "COMPANYNAME",
                 "COMPANY_ID",
@@ -148,7 +149,6 @@ HIERARCHY: dict = {
                 "HOSPITAL_NAME",
             ],
             "OTHER_ORG": [],
-            "ORG": [],
         },
         "EMPLOYMENT": {
             "JOB_TITLE": [